Test2

+++ app/Http/Controllers/MenuController.php
@@ -4,9 +4,9 @@ namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
 use App\Models\Menu;
-use App\Models\MenuIngredient;
 use Carbon\Carbon;
 use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Log;
 
 class MenuController extends Controller
 {
@@ -44,13 +44,13 @@ class MenuController extends Controller
 
     public function getRecipesForDate($date)
     {
-        // ログインしているユーザーのIDを取得
-        $userId = Auth::id();
+        // ログインしているユーザーを取得
+        $user = Auth::user();
 
         // 特定の日付に対応するログインしているユーザーのレシピの名前を取得
         $recipes = Menu::where('date', $date)
-            ->whereHas('dish', function ($query) use ($userId) {
-                $query->where('user_id', 1);
+            ->whereHas('dish', function ($query) use ($user) {
+                $query->where('user_id', $user->id);
             })
             ->with('dish') // 関連する料理情報も取得
             ->get();
@@ -58,17 +58,24 @@ class MenuController extends Controller
         return response()->json($recipes);
     }
 
+
+
     public function destroy($id)
     {
-        try {
-            $recipe = Menu::findOrFail($id);
-            $recipe->delete();
+        $recipe = Menu::find($id);
 
-            return response()->json([], 204); // 成功した場合は204 No Contentを返す
+        if (!$recipe) {
+            return response()->json(['error' => 'Not Found'], 404);
+        }
+
+        try {
+            $recipe->delete();
+            return response()->json([], 204);
         } catch (\Exception $e) {
             return response()->json(['error' => 'Internal Server Error'], 500);
         }
     }
+
 
     public function getIngredientsListData(Request $request)
     {
